refactor : ladder 패키지에 정리
