+++ src/main/java/src/model/ladder/Ladder.java
@@ -1,0 +1,13 @@
+package src.model.ladder;
+
+public class Ladder {
+	int[][] ladderBlueprint;
+
+	public Ladder(int[][] ladderBlueprint) {
+		this.ladderBlueprint = ladderBlueprint;
+	}
+
+	public int[][] getLadderBlueprint() {
+		return ladderBlueprint;
+	}
+}
