+++ src/test/java/src/model/LadderMakerTest.java
@@ -2,7 +2,8 @@ package src.model;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import src.model.ladder.Ladder;
+import src.model.ladder.LadderMaker;
 
 class LadderMakerTest {
 	@Test
